Make QCoreApplicationPrivate's destructor virtual in bootstrapped builds
In a bootstrapped build QCAP does not inherit from QObjectPrivate,
meaning it won't get a polymorphic destructor, causing compile errors.
This is necessary in preparation to add virtuals to QCAP.

The "virtual" keyword is protected to avoid further warnings by
compilers or linters telling that it's unnecessary because the base
class' dtor is already virtual.

Change-Id: Ifbd218154f74b4f1d2f2c5a3ef1fc31970f1c24b
Reviewed-by: David Faure <david.faure@kdab.com>
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>